Render skins in 0.7 demos
