+++ src/engine/shared/translation_context.h
@@ -1,6 +1,7 @@
 #ifndef ENGINE_SHARED_TRANSLATION_CONTEXT_H
 #define ENGINE_SHARED_TRANSLATION_CONTEXT_H
 
+#include <base/system.h>
 #include <engine/shared/protocol.h>
 #include <game/generated/protocol7.h>
 
@@ -98,6 +99,9 @@ public:
 
 	bool m_ShouldSendGameInfo;
 	int m_GameStateFlags7;
+	// 0.7 game flags
+	// use in combination with protocol7::GAMEFLAG_*
+	// for example protocol7::GAMEFLAG_TEAMS
 	int m_GameFlags;
 	int m_ScoreLimit;
 	int m_TimeLimit;
