+++ src/game/client/gameclient.h
@@ -365,12 +365,6 @@ public:
 		int m_Country;
 		char m_aSkinName[24];
 		int m_SkinColor;
-
-		// 0.7 Skin
-		char m_aaSkinPartNames[protocol7::NUM_SKINPARTS][protocol7::MAX_SKIN_LENGTH];
-		int m_aUseCustomColors[protocol7::NUM_SKINPARTS];
-		int m_aSkinPartColors[protocol7::NUM_SKINPARTS];
-
 		int m_Team;
 		int m_Emoticon;
 		float m_EmoticonStartFraction;
@@ -433,6 +427,17 @@ public:
 
 		void UpdateRenderInfo(bool IsTeamPlay);
 		void Reset();
+
+		class CSixup
+		{
+		public:
+			char m_aaSkinPartNames[protocol7::NUM_SKINPARTS][protocol7::MAX_SKIN_LENGTH];
+			int m_aUseCustomColors[protocol7::NUM_SKINPARTS];
+			int m_aSkinPartColors[protocol7::NUM_SKINPARTS];
+		};
+
+		// 0.7 Skin
+		CSixup m_Sixup;
 	};
 
 	CClientData m_aClients[MAX_CLIENTS];
@@ -493,7 +498,10 @@ public:
 	void OnStateChange(int NewState, int OldState) override;
 	template<typename T>
 	void ApplySkin7InfoFromGameMsg(const T *pMsg, int ClientId);
+	void ApplySkin7InfoFromSnapObj(const protocol7::CNetObj_De_ClientInfo *pObj, int ClientId) override;
+	int OnDemoRecSnap7(class CSnapshot *pFrom, class CSnapshot *pTo, int Conn) override;
 	void *TranslateGameMsg(int *pMsgId, CUnpacker *pUnpacker, int Conn);
+	int TranslateSnap(CSnapshot *pSnapDstSix, CSnapshot *pSnapSrcSeven, int Conn, bool Dummy) override;
 	void OnMessage(int MsgId, CUnpacker *pUnpacker, int Conn, bool Dummy) override;
 	void InvalidateSnapshot() override;
 	void OnNewSnapshot() override;
@@ -522,6 +530,7 @@ public:
 	const char *GetItemName(int Type) const override;
 	const char *Version() const override;
 	const char *NetVersion() const override;
+	const char *NetVersion7() const override;
 	int DDNetVersion() const override;
 	const char *DDNetVersionStr() const override;
 	virtual int ClientVersion7() const override;
